cs528-hw3-microservices/cloud function/requirements.txt
google-cloud-storage
google-cloud-pubsub
functions-framework
